requirements.txt
anyio==4.9.0
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.2
djangorestframework==3.16.0
ffmpeg==1.4
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
ngrok==1.4.0
packaging==24.2
python-telegram-bot==22.0
requests==2.32.3
sniffio==1.3.1
sqlparse==0.5.3
typing_extensions==4.13.1
tzdata==2025.2
urllib3==2.3.0
whitenoise==6.9.0
youtube-search-python==1.6.6
yt-dlp==2025.3.31
